refactor: Convert CookieUtils to object and set Access Token in header using HttpServletResponse
