+++ backend/chidic-app/src/main/kotlin/com/chidicapp/api/auth/OAuth2Controller.kt
@@ -1,8 +1,7 @@
 package com.chidicapp.api.auth
 
 import com.chidicapp.service.auth.AuthService
-import com.chidicapp.api.response.TokenResponse
-import com.chidiccore.util.CookieUtils
+import jakarta.servlet.http.Cookie
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
@@ -14,14 +13,11 @@ class OAuth2Controller(
 ) {
     @PostMapping("/login/kakao")
     @ResponseStatus(HttpStatus.OK)
-    fun login(@RequestHeader("X-Kakao-Token") kakaoToken: String, response: HttpServletResponse): TokenResponse {
+    fun login(@RequestHeader("X-Kakao-Token") kakaoToken: String, response: HttpServletResponse) {
         val tokenDto = authService.loginOrRegister(kakaoToken)
 
         CookieUtils.addRefreshTokenCookie(response, tokenDto.refreshToken)
-
-        return TokenResponse(
-            accessToken = tokenDto.accessToken
-        )
+        response.addHeader("Authorization", "Bearer ${tokenDto.accessToken}")
     }
 
     @PostMapping("/refresh")
@@ -29,14 +25,10 @@ class OAuth2Controller(
     fun refreshToken(
         @CookieValue("refresh_token") refreshToken: String,
         response: HttpServletResponse
-    ): TokenResponse {
+    ) {
         val tokenDto = authService.refreshAccessToken(refreshToken)
-
         CookieUtils.addRefreshTokenCookie(response, tokenDto.refreshToken)
-
-        return TokenResponse(
-            accessToken = tokenDto.accessToken
-        )
+        response.addHeader("Authorization", "Bearer ${tokenDto.accessToken}")
     }
 
     @PostMapping("/logout")
@@ -51,5 +43,25 @@ class OAuth2Controller(
         @PathVariable("userId") userId: Long
     ): String {
         return authService.makeTestAccessToken(userId)
+    }
+}
+
+object CookieUtils {
+    fun addRefreshTokenCookie(response: HttpServletResponse, refreshToken: String) {
+        val cookie = Cookie("refresh_token", refreshToken)
+        cookie.isHttpOnly = true
+        cookie.secure = true
+        cookie.path = "/api/refresh"
+        cookie.maxAge = 60 * 60 * 24 * 7
+        response.addCookie(cookie)
+    }
+
+    fun deleteRefreshTokenCookie(response: HttpServletResponse) {
+        val cookie = Cookie("refresh_token", "")
+        cookie.isHttpOnly = true
+        cookie.secure = true
+        cookie.path = "/api/refresh"
+        cookie.maxAge = 0
+        response.addCookie(cookie)
     }
 }
