--- backend/chidic-app/src/main/kotlin/com/chidicapp/api/response/TokenResponse.kt
@@ -1,5 +1,0 @@
-package com.chidicapp.api.response
-
-data class TokenResponse (
-    val accessToken: String
-)
