--- backend/chidic-core/src/main/kotlin/com/chidiccore/util/CookieUtils.kt
@@ -1,27 +1,0 @@
-package com.chidiccore.util
-
-import jakarta.servlet.http.Cookie
-import jakarta.servlet.http.HttpServletResponse
-
-class CookieUtils {
-
-    companion object{
-        fun addRefreshTokenCookie(response: HttpServletResponse, refreshToken: String) {
-            val cookie = Cookie("refresh_token", refreshToken)
-            cookie.isHttpOnly = true
-            cookie.secure = true
-            cookie.path = "/api/refresh"
-            cookie.maxAge = 60 * 60 * 24 * 7
-            response.addCookie(cookie)
-        }
-
-        fun deleteRefreshTokenCookie(response: HttpServletResponse) {
-            val cookie = Cookie("refresh_token", "")
-            cookie.isHttpOnly = true
-            cookie.secure = true
-            cookie.path = "/api/refresh"
-            cookie.maxAge = 0
-            response.addCookie(cookie)
-        }
-    }
-}
