+++ backend/chidic-domain/build.gradle.kts
@@ -15,7 +15,7 @@ dependencies {
 	implementation("org.mapstruct:mapstruct:1.5.5.Final")
 	kapt("org.mapstruct:mapstruct-processor:1.5.5.Final")
 	runtimeOnly("com.mysql:mysql-connector-j")
-	
+
 	// 테스트를 위한 의존성
 	testImplementation("org.springframework.boot:spring-boot-starter-test")
 	testImplementation("com.h2database:h2")
@@ -27,7 +27,6 @@ dependencies {
 	// 직렬화 라이브러리
 	implementation("com.fasterxml.jackson.core:jackson-databind:2.15.0")
 	implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.15.0")
-
 }
 
 allOpen {
